FEAT: modularize the `Result` field in actors
