+++ services/actors/search/types.ts
@@ -1,19 +1,8 @@
-import { KnownFor } from "services/actors/types/knowFor";
+import { GenericActor } from "services/actors/types";
 
 export interface ActorSearchResponse {
   page: number;
-  results: Result[];
+  results: GenericActor[];
   total_pages: number;
   total_results: number;
-}
-
-export interface Result {
-  adult: boolean;
-  gender: number;
-  id: number;
-  known_for: KnownFor[];
-  known_for_department: string;
-  name: string;
-  popularity: number;
-  profile_path: null | string;
 }
